style, pre-built images update

+++ src/utils.c
@@ -410,14 +410,14 @@ uint_fast8_t uint8_queue_empty(const u8queue_t * q)
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wshift-count-overflow"
 
-uint32_t ptr_hi32(uintptr_t v)
+uint_fast32_t ptr_hi32(uintptr_t v)
 {
 	if (sizeof v == sizeof (uint32_t))
 		return 0;
-	return v >> 32;
+	return (uint32_t) (v >> 32);
 }
 
-uint32_t ptr_lo32(uintptr_t v)
+uint_fast32_t ptr_lo32(uintptr_t v)
 {
 	return (uint32_t) v;
 }
